refactor: extract dev plugins from vite.config.ts into vite-plugins/
Move three inline Vite dev plugins into separate files to reduce
vite.config.ts complexity from 397 lines (health 8.38) to 77 lines:

- vite-plugins/vault-api.ts — vault API middleware (health 10.0)
- vite-plugins/ai-chat-proxy.ts — Anthropic proxy (health 10.0)
- vite-plugins/mcp-bridge.ts — MCP bridge info endpoint (health 10.0)

Also refactored vault-api internals: extracted route handlers into
individual functions with a dispatch table, and pulled frontmatter
helpers (getFmString, getFmArray, extractRelationships, parseDateToMs)
out of parseMarkdownFile to reduce cyclomatic complexity.
