+++ vite-plugins/mcp-bridge.ts
@@ -1,0 +1,18 @@
+import type { Plugin } from 'vite'
+
+/** WebSocket proxy info endpoint — tells the frontend where the MCP bridge is */
+export function mcpBridgeInfoPlugin(): Plugin {
+  return {
+    name: 'mcp-bridge-info',
+    configureServer(server) {
+      server.middlewares.use((req, res, next) => {
+        if (req.url !== '/api/mcp/info') return next()
+        res.setHeader('Content-Type', 'application/json')
+        res.end(JSON.stringify({
+          wsUrl: `ws://localhost:${process.env.MCP_WS_PORT || 9710}`,
+          available: true,
+        }))
+      })
+    },
+  }
+}
